prjcostprediction/pyproject.toml
[project]
name = "prjcostprediction"
version = "0.1.0"
description = "hugging face amazon data used for price prediction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "IPython",
    "accelerate>=0.24.1",
    "bitsandbytes>=0.41.3",
    "chromadb>=1.2.1",
    "datasets>=4.8.5",
    "gradio>=6.13.0",
    "huggingface-hub",
    "ipykernel",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-huggingface>=1.1.0",
    "langchain-ollama>=0.0.4",
    "langchain-openai>=1.2.1",
    "litellm>=1.83.0",
    "matplotlib>=3.10.9",
    "numpy>=1.24.0",
    "ollama>=0.6.2",
    "plotly>=5.14.3",
    "polars>=1.40.1",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.12",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=3.0.1",
    "torch>=2.0.0",
    "tqdm>=4.67.3",
    "transformers>=4.43.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cuda" }

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
skip-magic-trailing-comma = false
line-ending = "lf"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
